requirements.txt
# Project dependencies
telethon>=1.28.0
python-dotenv>=1.0.0
mindbase-layer @ git+https://github.com/aleksandr-dzhumurat/automation_toolkit.git
